requirements.txt
slackclient
python-ldap
aiohttp==3.6.0
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
hvac==0.9.5
idna==2.8
multidict==4.5.2
ply==3.11
pyasn1==0.4.7
pyasn1-modules==0.2.6
pyhcl==0.3.12
python-ldap==3.2.0
requests==2.22.0
slackclient==2.1.0
urllib3==1.25.3
yarl==1.3.0